+++ src/CHANGES.md
@@ -1,5 +1,6 @@
 **Added:**
-* none
+* about dialog
+* build number to splash screen
 
 **Changed:**
 * none